# docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.PolicyRequest:
    properties:
      description:
        type: string
      name:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.PolicyStepRequest'
        type: array
      team_id:
        description: nil = global
        type: string
    type: object
  api.PolicyStepRequest:
    properties:
      continue_on_failure:
        description: nil defaults to true
        type: boolean
      delay_seconds:
        type: integer
      max_attempts:
        type: integer
      message:
        type: string
      provider:
        description: '"slack", "telegram", ...'
        type: string
      target_id:
        type: string
      target_kind:
        description: '"dm" | "channel"'
        type: string
      target_type:
        description: user, channel, schedule
        type: string
      timeout_seconds:
        type: integer
    type: object
...
  model.EscalationPolicy:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      steps:
        items:
          $ref: '#/definitions/model.EscalationStep'
        type: array
      team_id:
        description: nil = global policy
        type: string
      updated_at:
        type: string
    type: object
...
  model.EscalationStep:
    properties:
      continue_on_failure:
        type: boolean
      delay_seconds:
        type: integer
      id:
        type: string
      max_attempts:
        type: integer
      message:
        type: string
      policy_id:
        type: string
      provider:
        description: '"slack", "telegram", ...'
        type: string
      step_index:
        type: integer
      target_id:
        description: actual ID
        type: string
      target_kind:
        description: '"dm" | "channel"'
        type: string
      target_type:
        description: user, channel, schedule
        type: string
      timeout_seconds:
        type: integer
    type: object
...
info:
  contact: {}
  description: TokayOps API
  title: TokayOps API
  version: "1.0"
paths:
...
  /api/v1/policies:
    get:
      description: Get all escalation policies (filtered by team membership for non-admins)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.EscalationPolicy'
            type: array
      summary: List all escalation policies
      tags:
      - policies
    post:
      consumes:
      - application/json
      description: Create a new escalation policy (admin only for global, team_admin
        for team-scoped)
      parameters:
      - description: Policy data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.EscalationPolicy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new escalation policy
      tags:
      - policies
  /api/v1/policies/{id}:
    delete:
      description: Delete escalation policy (admin for global, team_admin for team-scoped)
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete an escalation policy
      tags:
      - policies
    get:
      description: Get escalation policy by ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EscalationPolicy'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get a specific escalation policy
      tags:
      - policies
    put:
      consumes:
      - application/json
      description: Update escalation policy (admin for global, team_admin for team-scoped)
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Policy data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.EscalationPolicy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update an escalation policy
      tags:
      - policies
...
swagger: "2.0"